Hive adapter refactoring
