+++ src/main/java/com/exasol/adapter/HiveTableInformation.java
@@ -1,0 +1,4 @@
+package com.exasol.adapter;
+
+public class HiveTableInformation {
+}
